Remove useless placeholder
